+++ webapp/webapp/urls.py
@@ -23,8 +23,6 @@ urlpatterns = [
     path('accounts/', include('allauth.urls')),
     path('register/', views.register, name='register'),
     path('account/', views.account, name='account'),
-    path('create-checkout-session/', views.create_checkout_session, name='create-checkout-session'),
-    path('create-customer-portal-session/', views.create_customer_portal_session, name='create-customer-portal-session'),
     path('vocab-test/', views.vocab_test, name='vocab-test'),
     path('flag-question/', views.flag_question, name='flag-question'),
     path('vocab-builder/', views.vocab_builder, name='vocab-builder'),
